# perf: Avoid extra cloning of namespaces